requirements.txt
bcrypt
blinker
certifi
cffi
click
Flask
Flask-Bcrypt
Flask-Login
Flask-Mail
Flask-SQLAlchemy
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe
Pillow
pycparser==2.18
six==1.11.0
SQLAlchemy
Werkzeug==0.14.1
WTForms==2.1